+++ JointNode.cpp
@@ -14,14 +14,14 @@ JointNode::~JointNode() {
 
 }
  //---------------------------------------------------------------------------------------
-void JointNode::set_joint_x(double min, double init, double max) {
+void JointNode::set_joint_x(float min, float init, float max) {
 	m_joint_x.min = min;
 	m_joint_x.init = init;
 	m_joint_x.max = max;
 }
 
 //---------------------------------------------------------------------------------------
-void JointNode::set_joint_y(double min, double init, double max) {
+void JointNode::set_joint_y(float min, float init, float max) {
 	m_joint_y.min = min;
 	m_joint_y.init = init;
 	m_joint_y.max = max;
