+++ Light.hpp
@@ -12,7 +12,7 @@ struct Light {
   
   glm::vec3 colour;
   glm::vec3 position;
-  double falloff[3];
+  float falloff[3];
 };
 
 std::ostream& operator<<(std::ostream& out, const Light& l);
