+++ Ray.cpp
@@ -1,6 +1,6 @@
 #include "Ray.hpp"
 
-Ray::Ray(glm::dvec4 origin, glm::dvec4 vec) : origin(origin), vec(vec)
+Ray::Ray(glm::vec4 origin, glm::vec4 vec) : origin(origin), vec(vec)
 {}
 
 Ray::Ray() {}
